% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_empties}
\alias{drop_empties}
\title{Drop empty columns from a data.table}
\usage{
drop_empties(d)
}
\arguments{
\item{d}{\code{data.table}.}
}
\value{
\code{d} modified by reference, invisibly.
}
\description{
An empty column is one whose only values are \code{NA} or "".
}
\examples{
library('data.table')
d = data.table(w = 3:4, x = c('', 'foo'), y = c(NA, NA), z = c(NA, ''))
drop_empties(d)

}
