% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{exampleFunction}
\alias{exampleFunction}
\title{Example function and concise title}
\usage{
exampleFunction(a, b)
}
\arguments{
\item{a}{First numeric value to add.}

\item{b}{Second numeric value to add.}
}
\value{
A \code{data.table} with columns for \code{a}, \code{b}, and the result \code{r}.
}
\description{
Concise, non-redundant description of function.
}
\examples{
d = exampleFunction(3, 6)

}
