% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurveycto-package.R
\docType{package}
\name{rsurveycto-package}
\alias{rsurveycto}
\alias{rsurveycto-package}
\title{rsurveycto: Interact with Data on 'SurveyCTO'}
\description{
'SurveyCTO' is a platform for mobile data collection in offline settings. The 'rsurveycto' R package uses the 'SurveyCTO' REST API \url{https://docs.surveycto.com/05-exporting-and-publishing-data/05-api-access/01.api-access.html} to read datasets and forms from a 'SurveyCTO' server into R as 'data.table's and to download file attachments. The package also has limited support to write datasets to a server.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://agency-fund.github.io/rsurveycto/}
  \item \url{https://github.com/agency-fund/rsurveycto}
}

}
\author{
\strong{Maintainer}: Jake Hughey \email{jake@agency.fund}

Authors:
\itemize{
  \item Robert On \email{robert@agency.fund}
}

}
\keyword{internal}
