% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scto_get_form_definitions.R
\name{scto_get_form_definitions}
\alias{scto_get_form_definitions}
\title{Fetch form definitions from a SurveyCTO server}
\usage{
scto_get_form_definitions(auth, form_ids = NULL, deployed_only = FALSE)
}
\arguments{
\item{auth}{\code{\link[=scto_auth]{scto_auth()}} object.}

\item{form_ids}{Character vector indicating the form IDs. \code{NULL} indicates
all forms.}

\item{deployed_only}{Logical indicating whether to fetch definitions for all
versions of each form, or only for the deployed version.}
}
\value{
A \code{data.table} with one row per form (and per version, if
\code{deployed_only} is \code{FALSE}). Definitions are returned as nested
\code{data.table}s, which can be unnested using
\code{\link[=scto_unnest_form_definitions]{scto_unnest_form_definitions()}}.
}
\description{
This function fetches definitions for deployed and previous versions of one
or more forms.
}
\examples{
\dontrun{
auth = scto_auth('scto_auth.txt')
form_defs = scto_get_form_definitions(auth, 'my_form')
form_defs_unnest = scto_unnest_form_definitions(form_defs)
}

}
