% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scto_get_form_versions.R
\name{scto_get_form_versions}
\alias{scto_get_form_versions}
\title{Fetch form versions from a SurveyCTO server}
\usage{
scto_get_form_versions(auth, form_id, definitions = TRUE)
}
\arguments{
\item{auth}{\code{\link[=scto_auth]{scto_auth()}} object.}

\item{form_id}{String indicating the form ID.}

\item{definitions}{Logical indicating whether to fetch form definitions.}
}
\value{
A \code{data.table} with one row per version.
}
\description{
This function fetches a form's version history.
}
\examples{
\dontrun{
auth = scto_auth('scto_auth.txt')
form_versions = scto_get_form_versions(auth, 'my_form')
}

}
