% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scto_read.R
\name{scto_meta}
\alias{scto_meta}
\title{Access SurveyCTO meta data}
\usage{
scto_meta(auth)
}
\arguments{
\item{auth}{\code{\link[=scto_auth]{scto_auth()}} object.}
}
\value{
A list containing SurveyCTO server metadata
}
\description{
This function reads meta-data from SurveyCTO, specifically a list of
forms, datasets, groups and publishing information.
}
\examples{
\dontrun{
auth = scto_auth('scto_auth.txt')
scto_data = scto_meta(auth)
}

}
\seealso{
\code{\link[=scto_auth]{scto_auth()}}, \code{\link[=scto_read]{scto_read()}}, \code{\link[=scto_write]{scto_write()}}
}
