% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scto.R
\name{scto_pull}
\alias{scto_pull}
\title{Access SurveyCTO data using the API}
\usage{
scto_pull(
  auth,
  id,
  type = c("dataset", "form"),
  start_date = 1L,
  drop_empty_cols = TRUE,
  refresh = FALSE,
  cache_dir = "scto_data"
)
}
\arguments{
\item{auth}{\code{\link[=scto_auth]{scto_auth()}} object.}

\item{id}{String indicating ID of dataset or form to fetch.}

\item{type}{String indicating type of SurveyCTO data.}

\item{start_date}{Date, string coercible to a date, or an integer
(corresponding to days since 1970-01-01) indicating earliest date for which
to fetch data.}

\item{drop_empty_cols}{Logical indicating whether to drop columns that
contain only \code{NA} or only an empty string.}

\item{refresh}{Logical indicating whether to fetch fresh data using the API
or to attempt to load locally cached data.}

\item{cache_dir}{String indicating path to directory in which to cache data.}
}
\value{
A data.table.
}
\description{
This function pulls data from SurveyCTO using the API.
}
\examples{
\dontrun{
auth = scto_auth('scto_auth.txt')
test_data = scto_pull(auth, 'my_form', 'dataset')
}

}
\seealso{
\code{\link[=scto_push]{scto_push()}}
}
