% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scto.R
\name{scto_push}
\alias{scto_push}
\title{Upload data to SurveyCTO}
\usage{
scto_push(data, dataset_id, dataset_title, auth)
}
\arguments{
\item{data}{data.frame to upload.}

\item{dataset_id}{String indicating existing dataset ID on the server.}

\item{dataset_title}{String indicating title of dataset.}

\item{auth}{\code{\link[=scto_auth]{scto_auth()}} object.}
}
\value{
An object of class \code{\link[httr:response]{httr::response()}}.
}
\description{
This function uploads a csv file to SurveyCTO using web POSTs and GETs to
replace data in an existing Server Dataset.
}
\examples{
\dontrun{
auth = scto_auth('scto_auth.txt')
scto_push(data, 'my_dataset', 'My Dataset', auth)
}

}
\seealso{
\code{\link[=scto_pull]{scto_pull()}}
}
