% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scto_quiet}
\alias{scto_quiet}
\title{Suppress or permit messages from rsurveycto}
\usage{
scto_quiet(quiet = NULL)
}
\arguments{
\item{quiet}{A logical indicating whether to suppress messages, or \code{NULL}.}
}
\value{
If \code{quiet} is \code{NULL}, the current value of the \code{rsurveycto_quiet}
option. Otherwise, the previous value of the \code{rsurveycto_quiet} option
invisibly.
}
\description{
By default, rsurveycto prints messages to the console. To suppress them, set
the \code{rsurveycto_quiet} option to \code{TRUE} or use this function.
}
\examples{
options(rsurveycto_quiet = TRUE)
scto_quiet()
scto_quiet(FALSE)

}
