% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scto_get_form_versions.R
\name{scto_rbind_form_definitions}
\alias{scto_rbind_form_definitions}
\title{Combine form definitions from a SurveyCTO server}
\usage{
scto_rbind_form_definitions(form_versions)
}
\arguments{
\item{form_versions}{\code{data.table} returned by \code{\link[=scto_get_form_versions]{scto_get_form_versions()}}.}
}
\value{
A named list of \code{data.table}s for the survey, choices, and settings
components of the form definition.
}
\description{
This function combines form definitions, e.g., from multiple versions of a
form, which can make it easier to map values to labels in a later step.
}
\examples{
\dontrun{
auth = scto_auth('scto_auth.txt')
form_versions = scto_get_form_versions(auth, 'my_form')
form_defs_rbind = scto_rbind_form_definitions(form_versions)
}

}
