% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scto.R
\name{scto_read}
\alias{scto_read}
\title{Access SurveyCTO data using the API}
\usage{
scto_read(
  auth,
  id,
  type = c("dataset", "form"),
  start_date = "1900-01-01",
  review_status = "approved",
  drop_empty_cols = TRUE,
  refresh = TRUE,
  cache_dir = "scto_data"
)
}
\arguments{
\item{auth}{\code{\link[=scto_auth]{scto_auth()}} object.}

\item{id}{String indicating ID of the resource to fetch.}

\item{type}{String indicating whether the resource is a server dataset or a
form.}

\item{start_date}{Date-time or something coercible to a date-time
indicating the earliest date-time for which to fetch data. Only used for
forms.}

\item{review_status}{String or character vector indicating which submissions
to fetch. Possible values are "approved", "pending", "rejected", or any
combination of the three. Only used for forms.}

\item{drop_empty_cols}{Logical indicating whether to drop columns that
contain only \code{NA} or only an empty string.}

\item{refresh}{Logical indicating whether to fetch fresh data using the API
or to attempt to load locally cached data.}

\item{cache_dir}{String indicating path to directory in which to cache data.}
}
\value{
A \code{data.table}.
}
\description{
This function reads data from SurveyCTO using the API.
}
\examples{
\dontrun{
auth = scto_auth('scto_auth.txt')
test_data = scto_read(auth, 'my_form', 'dataset')
}

}
\seealso{
\code{\link[=scto_write]{scto_write()}}
}
