% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scto.R
\name{scto_upload}
\alias{scto_upload}
\title{Upload data to SurveyCTO}
\usage{
scto_upload(
  data,
  dataset_id,
  dataset_title,
  servername,
  auth_file = getOption("scto_auth_file")
)
}
\arguments{
\item{data}{data.frame to upload}

\item{dataset_id}{String indicating existing dataset ID on the server.}

\item{dataset_title}{String indicating title of dataset.}

\item{servername}{String indicating name of the SurveyCTO server.}

\item{auth_file}{String indicating path to file containing authorization
credentials, with the username on the first line and the password on the
second.}
}
\value{
An object of class \code{\link[httr:response]{httr::response()}}.
}
\description{
This function uploads a csv file to SurveyCTO using web POSTs and GETs to
replace data in an existing Server Dataset.
}
\examples{
\dontrun{
options(scto_auth_file = 'scto_auth.txt')
scto_upload(data, 'my_dataset', 'My Dataset', 'my_server')
}

}
