% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scto_get_attachments.R
\name{scto_get_attachments}
\alias{scto_get_attachments}
\title{Download file attachments from a SurveyCTO form}
\usage{
scto_get_attachments(
  auth,
  scto_data,
  column_name,
  output_dir = "scto_data",
  overwrite = TRUE
)
}
\arguments{
\item{auth}{\code{\link[=scto_auth]{scto_auth()}} object.}

\item{scto_data}{\code{data.table} from \code{\link[=scto_read]{scto_read()}} containing data from a
SurveyCTO form.}

\item{column_name}{String indicating column in \code{scto_data} for which to
download file attachments. Should contain URLs, can contain missing values.}

\item{output_dir}{String indicating path to directory in which to save files.}

\item{overwrite}{Logical indicating whether to overwrite existing files.}
}
\value{
A character vector of file names. Each element corresponds to a row
of \code{scto_data}, with \code{NA} for rows lacking a valid URL.
}
\description{
This function downloads files in bulk.
}
\examples{
\dontrun{
auth = scto_auth('scto_auth.txt')
scto_data = scto_read(auth, 'my_form', 'form')
filenames = scto_get_attachments(auth, scto_data, 'my_column')
}

}
\seealso{
\code{\link[=scto_auth]{scto_auth()}}, \code{\link[=scto_read]{scto_read()}}, \code{\link[=scto_write]{scto_write()}}
}
