% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scto_get_form_versions.R
\name{scto_get_form_versions}
\alias{scto_get_form_versions}
\title{Fetch form versions from a SurveyCTO server}
\usage{
scto_get_form_versions(auth, form_ids = NULL, get_defs = !is.null(form_ids))
}
\arguments{
\item{auth}{\code{\link[=scto_auth]{scto_auth()}} object.}

\item{form_ids}{Character vector indicating the form IDs. \code{NULL} indicates
all forms.}

\item{get_defs}{Logical indicating whether to fetch form definitions.}
}
\value{
A \code{data.table} with one row per version per form. Definitions will be
returned as nested \code{data.table}s, which can be unnested using
\code{\link[=scto_unnest_form_definitions]{scto_unnest_form_definitions()}}.
}
\description{
This function fetches a form's version history.
}
\examples{
\dontrun{
auth = scto_auth('scto_auth.txt')
form_versions = scto_get_form_versions(auth, 'my_form')
form_defs = scto_unnest_form_definitions(form_versions)
}

}
