% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scto.R
\name{scto_write}
\alias{scto_write}
\title{Upload data to SurveyCTO}
\usage{
scto_write(auth, data, dataset_id, dataset_title)
}
\arguments{
\item{auth}{\code{\link[=scto_auth]{scto_auth()}} object.}

\item{data}{\code{data.frame} to upload.}

\item{dataset_id}{String indicating ID of existing server dataset.}

\item{dataset_title}{String indicating title of dataset.}
}
\value{
An object of class \code{\link[httr:response]{httr::response()}}.
}
\description{
This function uploads data to SurveyCTO using web POSTs and GETs to
replace an existing server dataset.
}
\examples{
\dontrun{
auth = scto_auth('scto_auth.txt')
res = scto_write(auth, data, 'my_dataset', 'My Dataset')
}

}
\seealso{
\code{\link[=scto_read]{scto_read()}}
}
